Expose HAB_FALLBACK_CHANNEL in plan build
This was an internal variable before; this merely exposes it
externally in case it should ever be needed.

It is not currently intended that anyone outside of Chef would use
this, and thus will not be formally documented. It is being introduced
as an aid for helping to evaluate core plans refresh
candidates (specifically, to help our existing release pipeline to
build our Habitat packages with dependencies from a refresh channel,
rather than `stable`).

Signed-off-by: Christopher Maier <cmaier@chef.io>

fixup! Expose HAB_FALLBACK_CHANNEL in hab-plan-build

Signed-off-by: Christopher Maier <cmaier@chef.io>