[CI] Modify release pipeline to help evaluate a core plans refresh
In order to easily build a consistent set of packages that depend on
core plans from a refresh candidate channel, we need to tweak the
release pipeline a bit.

To do so, one would manually trigger a release pipeline run, setting
the following environment variables:

- HAB_FALLBACK_CHANNEL=<NAME_OF_REFRESH_CHANNEL>
- COLLECTION_CHANNEL=<SOMETHING_UNIQUE_AND_MEANINGFUL_TO_YOU>

This will build a set of packages and promote them into the
`$COLLECTION_CHANNEL` for subsequent evaluation and testing.

Normal, Expeditor-triggered pipeline runs will not have either of
these variables set, and will thus behave "normally".

Note that if `$COLLECTION_CHANNEL` is not set, the packages will still
be built, but will not be collected anywhere convenient once the
pipeline execution finishes.

Signed-off-by: Christopher Maier <cmaier@chef.io>